GA_6/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub>=2.15.1
google-cloud-storage>=2.7.0